python-tests/requirements.txt
json-logic-qubit>=0.9.0
panzi-json-logic>=1.0.1
